requirements.txt
install==1.3.5
Pillow==9.1.0
PyYAML==6.0
